requirements.txt
selenium
pytest
pillow
openai
allure-pytest
